+++ src/RcsbFv/RcsbFvBoard/RcsbFvBoard.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import {RcsbFvBoardConfigInterface, RcsbFvRowConfigInterface} from "../RcsbFvConfig/RcsbFvConfigInterface";
+import {RcsbFvBoardConfigInterface} from "../RcsbFvConfig/RcsbFvConfigInterface";
 import classes from "../RcsbFvStyles/RcsbFvRow.module.scss";
 
 import {
@@ -11,22 +11,22 @@ import {Subscription} from "rxjs";
 import {RcsbSelection} from "../../RcsbBoard/RcsbSelection";
 import {RcsbFvUI} from "../RcsbFvUI/RcsbFvUI";
 import {RcsbFvDOMConstants} from "../RcsbFvConfig/RcsbFvDOMConstants";
-import uniqid from "uniqid";
 import {RcsbFvTable} from "./RcsbFvTable";
 import {BoardGlow} from "./Components/BoardGlow";
 import {RowGlow} from "./Components/RowGlow";
-import {BoardProgress} from "./Components/BoardProgress";
-import {RowStatusMap} from "./Utils/RowStatusMap";
 import {RcsbScaleInterface} from "../../RcsbBoard/RcsbD3/RcsbD3ScaleFactory";
+import {RcsbFvExtendedRowConfigInterface} from "./Utils/BoardDataState";
+import {BoardProgress} from "./Components/BoardProgress";
 
 /**Board React component configuration interface*/
 export interface RcsbFvBoardFullConfigInterface {
-    readonly rowConfigData: Array<RcsbFvRowConfigInterface & {key:string}>;
+    readonly rowConfigData: Array<RcsbFvExtendedRowConfigInterface>;
     readonly boardConfigData: RcsbFvBoardConfigInterface;
 }
 
 /**Board React component interface*/
 interface RcsbFvBoardInterface extends RcsbFvBoardFullConfigInterface {
+    readonly boardId: string;
     readonly contextManager: RcsbFvContextManager;
     readonly resolve: ()=> void;
     readonly xScale: RcsbScaleInterface;
@@ -35,7 +35,7 @@ interface RcsbFvBoardInterface extends RcsbFvBoardFullConfigInterface {
 
 /**Board React component state interface*/
 interface RcsbFvBoardState {
-    readonly rowConfigData: Array<RcsbFvRowConfigInterface & {key:string}>;
+    readonly rowConfigData: Array<RcsbFvExtendedRowConfigInterface>;
     readonly boardConfigData: RcsbFvBoardConfigInterface;
     readonly progressStatus: number;
 }
@@ -43,8 +43,6 @@ interface RcsbFvBoardState {
 /**Board React Component className*/
 export class RcsbFvBoard extends React.Component <RcsbFvBoardInterface, RcsbFvBoardState > {
 
-    /**Inner div board DOM element id*/
-    private readonly boardId : string = uniqid("RcsbFvBoard_");
     /**Subscription to events*/
     private subscription: Subscription;
     /**Global d3 Xscale object shaed among all board tracks*/
@@ -53,8 +51,6 @@ export class RcsbFvBoard extends React.Component <RcsbFvBoardInterface, RcsbFvBo
     private readonly selection:RcsbSelection;
     /**Promise resolve callback when board is complete*/
     private resolveOnReady: (()=>void) | undefined = undefined;
-
-    private readonly rowStatusMap: RowStatusMap = new RowStatusMap();
 
 
     readonly state : RcsbFvBoardState = {
@@ -78,23 +74,21 @@ export class RcsbFvBoard extends React.Component <RcsbFvBoardInterface, RcsbFvBo
         return (
             <div className={classes.rcsbFvRootContainer} onMouseOver={this.setMouseOverCallback()} onMouseLeave={this.setMouseLeaveCallback()}>
                 {
-                    this.state.boardConfigData.disableMenu ? null : <RcsbFvUI boardId={this.boardId} boardConfigData={this.state.boardConfigData} contextManager={this.props.contextManager} xScale={this.xScale} />
+                    this.state.boardConfigData.disableMenu ? null : <RcsbFvUI boardId={this.props.boardId} boardConfigData={this.state.boardConfigData} contextManager={this.props.contextManager} xScale={this.xScale} />
                 }
-                <BoardGlow boardId={this.boardId} boardConfigData={this.state.boardConfigData} contextManager={this.props.contextManager}/>
-                <RowGlow boardId={this.boardId} boardConfigData={this.state.boardConfigData} contextManager={this.props.contextManager}/>
+                <BoardGlow boardId={this.props.boardId} boardConfigData={this.state.boardConfigData} contextManager={this.props.contextManager}/>
+                <RowGlow boardId={this.props.boardId} boardConfigData={this.state.boardConfigData} contextManager={this.props.contextManager}/>
                 <RcsbFvTable
-                    boardId={this.boardId}
+                    boardId={this.props.boardId}
                     xScale={this.xScale}
                     selection={this.selection}
                     contextManager={this.props.contextManager}
                     boardConfigData={this.state.boardConfigData}
                     rowConfigData={this.state.rowConfigData}
-                    resolve={this.props.resolve}
-                    rowStatusMap={this.rowStatusMap}
                 />
-                <div id={this.boardId+RcsbFvDOMConstants.TOOLTIP_DOM_ID_PREFIX} className={classes.rcsbFvTooltip} {...{[RcsbFvDOMConstants.POPPER_HIDDEN]:""}} />
-                <div id={this.boardId+RcsbFvDOMConstants.TOOLTIP_DESCRIPTION_DOM_ID_PREFIX} className={classes.rcsbFvTooltipDescription} {...{[RcsbFvDOMConstants.POPPER_HIDDEN]:""}} />
-                <BoardProgress boardId={this.boardId} boardConfigData={this.state.boardConfigData} rowConfigData={this.state.rowConfigData} contextManager={this.props.contextManager} rowStatusMap={this.rowStatusMap}/>
+                <div id={this.props.boardId+RcsbFvDOMConstants.TOOLTIP_DOM_ID_PREFIX} className={classes.rcsbFvTooltip} {...{[RcsbFvDOMConstants.POPPER_HIDDEN]:""}} />
+                <div id={this.props.boardId+RcsbFvDOMConstants.TOOLTIP_DESCRIPTION_DOM_ID_PREFIX} className={classes.rcsbFvTooltipDescription} {...{[RcsbFvDOMConstants.POPPER_HIDDEN]:""}} />
+                <BoardProgress boardId={this.props.boardId} boardConfigData={this.state.boardConfigData} rowConfigData={this.state.rowConfigData} contextManager={this.props.contextManager} />
             </div>
         );
     }
@@ -109,13 +103,13 @@ export class RcsbFvBoard extends React.Component <RcsbFvBoardInterface, RcsbFvBo
     }
 
     componentDidUpdate(prevProps: Readonly<RcsbFvBoardInterface>, prevState: Readonly<RcsbFvBoardState>, snapshot?: any) {
-        this.checkReadyState();
+        this.checkReadyState(prevState);
     }
 
     /**Subscribe className to rxjs events (adding tracks, change scale, update board config)
      * @return rxjs Subscription object
      * */
-    private subscribe(): Subscription{
+    private subscribe(): Subscription {
         return this.props.contextManager.subscribe((obj:RcsbFvContextManagerType)=>{
             switch (obj.eventType){
                 case EventType.UPDATE_BOARD_CONFIG:
@@ -124,9 +118,6 @@ export class RcsbFvBoard extends React.Component <RcsbFvBoardInterface, RcsbFvBo
                     break;
                 case EventType.BOARD_READY:
                     this.boardReady();
-                    break;
-                case EventType.UPDATE_BOARD_DATA:
-                    this.resetReadyStatus(obj.eventResolve);
                     break;
             }
         });
@@ -137,11 +128,24 @@ export class RcsbFvBoard extends React.Component <RcsbFvBoardInterface, RcsbFvBo
      * */
     private updateBoardConfig(configData: Partial<RcsbFvBoardFullConfigInterface>): void {
         if(configData.rowConfigData!=null && configData.boardConfigData!=null){
-            this.setState({rowConfigData: configData.rowConfigData, boardConfigData: {...this.state.boardConfigData, ...configData.boardConfigData}});
+            this.setState({
+                rowConfigData: configData.rowConfigData,
+                boardConfigData: {
+                    ...this.state.boardConfigData,
+                    ...configData.boardConfigData
+                }
+            });
         }else if(configData.boardConfigData!=null){
-            this.setState({boardConfigData: {...this.state.boardConfigData, ...configData.boardConfigData}} );
+            this.setState({
+                boardConfigData: {
+                    ...this.state.boardConfigData,
+                    ...configData.boardConfigData
+                }
+            });
         }else if(configData.rowConfigData!=null){
-            this.setState({rowConfigData: configData.rowConfigData} );
+            this.setState({
+                rowConfigData: configData.rowConfigData
+            });
         }
     }
 
@@ -159,29 +163,26 @@ export class RcsbFvBoard extends React.Component <RcsbFvBoardInterface, RcsbFvBo
             return this.onMouseLeave.bind(this);
     }
 
-    private onMouseOver (): void{
+    private onMouseOver(): void{
         this.boardHover(true);
     }
 
-    private onMouseLeave (): void{
+    private onMouseLeave(): void{
         this.boardHover(false);
     }
 
-    private resetReadyStatus(resolve: (()=>void) | undefined): void {
-        this.rowStatusMap.clear();
-        this.resolveOnReady = resolve;
+    private resetReadyStatus( eventResolve?: ()=>void): void {
+        this.resolveOnReady = eventResolve;
     }
 
-    /**Trigger Board Ready Event
-     * */
-    private boardReady():void{
-        if(typeof this.resolveOnReady === "function")
-            this.resolveOnReady();
-    }
-
-    private checkReadyState(): void {
-        if(!this.state.rowConfigData || this.state.rowConfigData.length == 0)
+    private checkReadyState(prevState?: Readonly<RcsbFvBoardState>): void {
+        if(
+            ( !this.state.rowConfigData ) ||
+            ( this.state.rowConfigData.length == 0 ) ||
+            ( prevState && this.state.rowConfigData.map(r=>r.key).sort().join("-") == prevState.rowConfigData.map(r=>r.key).sort().join("-") )
+        ){
             this.boardReady();
+        }
     }
 
     private renderStarts(): void {
@@ -192,6 +193,16 @@ export class RcsbFvBoard extends React.Component <RcsbFvBoardInterface, RcsbFvBo
         if(typeof this.state.boardConfigData.selectionChangeCallBack === "function")
             this.selection.setSelectionChangeCallback(this.state.boardConfigData.selectionChangeCallBack);
 
+    }
+
+    /**
+     * Trigger Board Ready Event
+     **/
+    private boardReady(): void{
+        if(typeof this.resolveOnReady === "function") {
+            this.resolveOnReady();
+            this.resolveOnReady = undefined;
+        }
     }
 
     private boardHover(flag:boolean): void{