+++ src/RcsbFv/RcsbFvBoard/Utils/BoardDataState.ts
@@ -1,34 +1,53 @@
 import {RcsbFvDisplayConfigInterface, RcsbFvRowConfigInterface} from "../../RcsbFvConfig/RcsbFvConfigInterface";
 import uniqid from "uniqid";
 import {
+    EventType,
+    RcsbFvContextManager,
+    RowReadyInterface,
     TrackDataInterface,
     TrackVisibilityInterface
 } from "../../RcsbFvContextManager/RcsbFvContextManager";
 import {RcsbFvDisplayTypes} from "../../RcsbFvConfig/RcsbFvDefaultConfigValues";
 import {arrayMoveMutable} from "array-move";
+import {RowStatusMap} from "./RowStatusMap";
+import {Subscription} from "rxjs";
+
+export interface RcsbFvExtendedRowConfigInterface extends RcsbFvRowConfigInterface {
+    key:string;
+    renderSchedule?: "async"|"sync"|"fixed";
+}
 
 export class BoardDataState {
 
-    private rowConfigData: (RcsbFvRowConfigInterface & {key:string})[] = []
+    private rowConfigData: RcsbFvExtendedRowConfigInterface[] = [];
+    private readonly rowStatusMap: RowStatusMap;
+    private readonly contextManager: RcsbFvContextManager;
+    private readonly subscription: Subscription;
 
-    constructor(rowConfigData?: RcsbFvRowConfigInterface[]) {
+    constructor(rowStatusMap: RowStatusMap, contextManager: RcsbFvContextManager, rowConfigData?: RcsbFvRowConfigInterface[]) {
+        this.rowStatusMap = rowStatusMap;
+        this.contextManager = contextManager;
         if(rowConfigData)
-            this.rowConfigData = rowConfigData.map(r=>BoardDataState.checkRow(r));
+            this.rowConfigData = rowConfigData.map(r=>this.checkRow(r));
+       this.subscription = this.subscribe();
     }
 
-    public getBoardData(): (RcsbFvRowConfigInterface & {key:string})[] {
+    public getBoardData(): RcsbFvExtendedRowConfigInterface[] {
         return this.rowConfigData;
     }
 
     public setBoardData(rowConfigData: RcsbFvRowConfigInterface[]): void {
-        this.rowConfigData  = rowConfigData.map(r=>BoardDataState.checkRow(r));
+        this.rowConfigData  = rowConfigData.map(r=>this.checkRow(r));
     }
 
     /**Adds a new track to the board
      * @param configRow Track configuration object
      * */
     public addTrack(configRow: RcsbFvRowConfigInterface): void{
-        this.rowConfigData.push( BoardDataState.checkRow(configRow) );
+        this.rowConfigData.push( {
+            ...this.checkRow(configRow),
+            renderSchedule: "sync"
+        } );
     }
 
     /**Modifies visibility of a board track
@@ -40,7 +59,6 @@ export class BoardDataState {
             return;
         row.trackVisibility = obj.visibility;
         row.key = BoardDataState.generateKey(row);
-
     }
 
     public moveTrack(move: {oldIndex: number, newIndex: number}): void {
@@ -69,6 +87,11 @@ export class BoardDataState {
         if(!row)
             return;
         row.key = BoardDataState.generateKey(row);
+        row.renderSchedule = "sync";
+    }
+
+    public unsubscribe(): void {
+        this.subscription.unsubscribe();
     }
 
     /**Modifies a board track data
@@ -95,11 +118,13 @@ export class BoardDataState {
             });
         }
         row.key = BoardDataState.generateKey(row);
+        row.renderSchedule = "sync"
     }
 
-    private static checkRow(d: RcsbFvRowConfigInterface): RcsbFvRowConfigInterface & {key:string} {
+    private checkRow(d: RcsbFvRowConfigInterface): RcsbFvRowConfigInterface & {key:string} {
         d.trackId = d.trackId ?? uniqid("trackId_");
         d.trackVisibility = typeof d.trackVisibility != "boolean" ? d.trackVisibility : true;
+        this.rowStatusMap.set(d.trackId,false);
         return {
             ...d,
             key: BoardDataState.generateKey(d)
@@ -108,6 +133,37 @@ export class BoardDataState {
 
     private static generateKey(d: RcsbFvRowConfigInterface): string {
         return `${d.trackId}_${uniqid("key_")}`;
+    }
+
+     private subscribe(): Subscription {
+         return  this.contextManager.subscribe((o)=>{
+            switch (o.eventType){
+                case EventType.ROW_READY:
+                    this.rowReady(o.eventData);
+                    break;
+            }
+        });
+    }
+
+    /**Row Track Board Ready Event
+     * @param rowData
+     * */
+    private rowReady(rowData:RowReadyInterface):void{
+        this.rowStatusMap.set(rowData.rowId,true);
+        this.contextManager.next({
+            eventType: EventType.FRACTION_COMPLETED,
+            eventData: Math.ceil(this.rowStatusMap.completed()/this.rowStatusMap.size()*100)
+        })
+        if(this.rowStatusMap.complete()){
+            this.boardReady();
+        }
+    }
+
+    private boardReady(){
+        this.contextManager.next({
+            eventType: EventType.BOARD_READY,
+            eventData: null
+        })
     }
 
 }