+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 [Semantic Versioning](https://semver.org/)
 
+## [2.5.0] - 2022-11-18
+### Code refactoring
+- All board config/data update events use `EventType.UPDATE_BOARD_CONFIG` type
+- Class `RcsbFvStateManager` handles zoom/move and selection events
+
 ## [2.4.0] - 2022-11-16
 ### Improvement
 - New class `BoardDataState` to handle the state of the board tracks data. It encodes the logic of changing or updating